+++ api/trading_test.go
@@ -285,6 +285,8 @@ func getTestGRPCServer(
 	sqlDepositStore := sqlstore.NewDeposits(&sqlStore)
 	sqlProposalStore := sqlstore.NewProposals(&sqlStore)
 	sqlVoteStore := sqlstore.NewVotes(&sqlStore)
+	sqlRiskFactorsStore := sqlstore.NewRiskFactors(&sqlStore)
+	sqlMarginLevelsStore := sqlstore.NewMarginLevels(&sqlStore)
 
 	g := api.NewGRPCServer(
 		logger,
@@ -331,6 +333,8 @@ func getTestGRPCServer(
 		sqlDepositStore,
 		sqlProposalStore,
 		sqlVoteStore,
+		sqlRiskFactorsStore,
+		sqlMarginLevelsStore,
 	)
 	if g == nil {
 		err = fmt.Errorf("failed to create gRPC server")