+++ cmd/data-node/node/node.go
@@ -117,6 +117,8 @@ type NodeCommand struct {
 	depositStoreSQL       *sqlstore.Deposits
 	proposalStoreSQL      *sqlstore.Proposals
 	voteStoreSQL          *sqlstore.Votes
+	marginLevelsStoreSQL  *sqlstore.MarginLevels
+	riskFactorStoreSQL    *sqlstore.RiskFactors
 
 	vegaCoreServiceClient vegaprotoapi.CoreServiceClient
 
@@ -161,6 +163,8 @@ type NodeCommand struct {
 	depositSubSQL          *sqlsubscribers.Deposit
 	proposalsSubSQL        *sqlsubscribers.Proposal
 	votesSubSQL            *sqlsubscribers.Vote
+	marginLevelsSubSQL     *sqlsubscribers.MarginLevels
+	riskFactorSubSQL       *sqlsubscribers.RiskFactor
 
 	candleService     *candles.Svc
 	tradeService      *trades.Svc
@@ -357,6 +361,8 @@ func (l *NodeCommand) createGRPCServer(config api.Config, useSQLStores bool) *ap
 		l.depositStoreSQL,
 		l.proposalStoreSQL,
 		l.voteStoreSQL,
+		l.riskFactorStoreSQL,
+		l.marginLevelsStoreSQL,
 	)
 	return grpcServer
 }