+++ cmd/data-node/node/node_pre.go
@@ -138,6 +138,8 @@ func (l *NodeCommand) setupSQLSubscribers() {
 	l.depositSubSQL = sqlsubscribers.NewDeposit(l.depositStoreSQL, l.Log)
 	l.proposalsSubSQL = sqlsubscribers.NewProposal(l.proposalStoreSQL, l.Log)
 	l.votesSubSQL = sqlsubscribers.NewVote(l.voteStoreSQL, l.Log)
+	l.marginLevelsSubSQL = sqlsubscribers.NewMarginLevels(l.marginLevelsStoreSQL, l.Log)
+	l.riskFactorSubSQL = sqlsubscribers.NewRiskFactor(l.riskFactorStoreSQL, l.Log)
 }
 
 func (l *NodeCommand) setupStorages() error {
@@ -180,6 +182,9 @@ func (l *NodeCommand) setupStorages() error {
 		l.depositStoreSQL = sqlstore.NewDeposits(sqlStore)
 		l.proposalStoreSQL = sqlstore.NewProposals(sqlStore)
 		l.voteStoreSQL = sqlstore.NewVotes(sqlStore)
+		l.marginLevelsStoreSQL = sqlstore.NewMarginLevels(sqlStore)
+		l.riskFactorStoreSQL = sqlstore.NewRiskFactors(sqlStore)
+
 		l.sqlStore = sqlStore
 	}
 
@@ -265,7 +270,9 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 			l.depositSubSQL,
 			l.proposalsSubSQL,
 			l.votesSubSQL,
-		)
+			l.depositSubSQL,
+			l.marginLevelsSubSQL,
+			l.riskFactorSubSQL)
 	}
 
 	l.broker, err = broker.New(l.ctx, l.Log, l.conf.Broker, l.chainInfoStore, eventSource)