+++ sqlstore/margin_level_test.go
@@ -1,0 +1,341 @@
+package sqlstore_test
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"code.vegaprotocol.io/data-node/entities"
+	"code.vegaprotocol.io/data-node/sqlstore"
+	"code.vegaprotocol.io/protos/vega"
+	"github.com/jackc/pgx/v4"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMarginLevels(t *testing.T) {
+	t.Run("Upsert should insert margin levels that don't exist in the current block", testInsertMarginLevels)
+	t.Run("Upsert should update margin levels that already exist in the same block", testUpdateDuplicateMarginLevelInSameBlock)
+	t.Run("GetMarginLevelsByID should return the latest state of margin levels for all markets if only party ID is provided", testGetMarginLevelsByPartyID)
+	t.Run("GetMarginLevelsByID should return the latest state of margin levels for all parties if only market ID is provided", testGetMarginLevelsByMarketID)
+	t.Run("GetMarginLevelsByID should return the latest state of margin levels for the given party/market ID", testGetMarginLevelsByID)
+}
+
+func setupMarginLevelTests(t *testing.T, ctx context.Context) (*sqlstore.Blocks, *sqlstore.MarginLevels, *pgx.Conn) {
+	t.Helper()
+	err := testStore.DeleteEverything()
+	require.NoError(t, err)
+
+	bs := sqlstore.NewBlocks(testStore)
+	ml := sqlstore.NewMarginLevels(testStore)
+	config := NewTestConfig(testDBPort)
+
+	conn, err := pgx.Connect(ctx, connectionString(config))
+	require.NoError(t, err)
+
+	return bs, ml, conn
+}
+
+func testInsertMarginLevels(t *testing.T) {
+	testTimeout := time.Second * 10
+	ctx, cancel := context.WithTimeout(context.Background(), testTimeout)
+	defer cancel()
+
+	bs, ml, conn := setupMarginLevelTests(t, ctx)
+
+	var rowCount int
+	err := conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+
+	block := addTestBlock(t, bs)
+	marginLevelProto := getMarginLevelProto()
+	marginLevel, err := entities.MarginLevelsFromProto(marginLevelProto, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 1, rowCount)
+}
+
+func testUpdateDuplicateMarginLevelInSameBlock(t *testing.T) {
+	testTimeout := time.Second * 10
+	ctx, cancel := context.WithTimeout(context.Background(), testTimeout)
+	defer cancel()
+
+	bs, ml, conn := setupMarginLevelTests(t, ctx)
+
+	var rowCount int
+	err := conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+
+	block := addTestBlock(t, bs)
+	marginLevelProto := getMarginLevelProto()
+	marginLevel, err := entities.MarginLevelsFromProto(marginLevelProto, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 1, rowCount)
+
+	err = ml.Upsert(marginLevel)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 1, rowCount)
+}
+
+func getMarginLevelProto() *vega.MarginLevels {
+	return &vega.MarginLevels{
+		MaintenanceMargin:      "1000",
+		SearchLevel:            "1000",
+		InitialMargin:          "1000",
+		CollateralReleaseLevel: "1000",
+		PartyId:                "DEADBEEF",
+		MarketId:               "DEADBEEF",
+		Asset:                  "DEADBEEF",
+		Timestamp:              time.Now().UnixNano(),
+	}
+}
+
+func testGetMarginLevelsByPartyID(t *testing.T) {
+	testTimeout := time.Second * 10
+	ctx, cancel := context.WithTimeout(context.Background(), testTimeout)
+	defer cancel()
+
+	bs, ml, conn := setupMarginLevelTests(t, ctx)
+
+	var rowCount int
+	err := conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 0, rowCount)
+
+	ml1 := getMarginLevelProto()
+	ml2 := getMarginLevelProto()
+	ml3 := getMarginLevelProto()
+	ml4 := getMarginLevelProto()
+
+	ml2.MarketId = "DEADBAAD"
+
+	ml3.Timestamp = ml2.Timestamp + 1000000000
+	ml3.MaintenanceMargin = "2000"
+	ml3.SearchLevel = "2000"
+
+	ml4.Timestamp = ml2.Timestamp + 1000000000
+	ml4.MaintenanceMargin = "2000"
+	ml4.SearchLevel = "2000"
+	ml4.MarketId = "DEADBAAD"
+
+	block := addTestBlock(t, bs)
+	marginLevel1, err := entities.MarginLevelsFromProto(ml1, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel2, err := entities.MarginLevelsFromProto(ml2, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel1)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel2)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 2, rowCount)
+
+	time.Sleep(time.Second)
+
+	block = addTestBlock(t, bs)
+	marginLevel3, err := entities.MarginLevelsFromProto(ml3, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel4, err := entities.MarginLevelsFromProto(ml4, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel3)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel4)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 4, rowCount)
+
+	got, err := ml.GetMarginLevelsByID(ctx, "DEADBEEF", "", entities.Pagination{})
+	assert.NoError(t, err)
+	assert.Equal(t, 2, len(got))
+
+	// We have to truncate the time because Postgres only supports time to microsecond granularity.
+	want1 := *marginLevel3
+	want1.Timestamp = want1.Timestamp.Truncate(time.Microsecond)
+	want1.VegaTime = want1.VegaTime.Truncate(time.Microsecond)
+
+	want2 := *marginLevel4
+	want2.Timestamp = want2.Timestamp.Truncate(time.Microsecond)
+	want2.VegaTime = want2.VegaTime.Truncate(time.Microsecond)
+
+	want := []entities.MarginLevels{want1, want2}
+
+	assert.ElementsMatch(t, want, got)
+}
+
+func testGetMarginLevelsByMarketID(t *testing.T) {
+	testTimeout := time.Second * 10
+	ctx, cancel := context.WithTimeout(context.Background(), testTimeout)
+	defer cancel()
+
+	bs, ml, conn := setupMarginLevelTests(t, ctx)
+
+	var rowCount int
+	err := conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 0, rowCount)
+
+	ml1 := getMarginLevelProto()
+	ml2 := getMarginLevelProto()
+	ml3 := getMarginLevelProto()
+	ml4 := getMarginLevelProto()
+
+	ml2.PartyId = "DEADBAAD"
+
+	ml3.Timestamp = ml2.Timestamp + 1000000000
+	ml3.MaintenanceMargin = "2000"
+	ml3.SearchLevel = "2000"
+
+	ml4.Timestamp = ml2.Timestamp + 1000000000
+	ml4.MaintenanceMargin = "2000"
+	ml4.SearchLevel = "2000"
+	ml4.PartyId = "DEADBAAD"
+
+	block := addTestBlock(t, bs)
+	marginLevel1, err := entities.MarginLevelsFromProto(ml1, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel2, err := entities.MarginLevelsFromProto(ml2, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel1)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel2)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 2, rowCount)
+
+	time.Sleep(time.Second)
+
+	block = addTestBlock(t, bs)
+	marginLevel3, err := entities.MarginLevelsFromProto(ml3, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel4, err := entities.MarginLevelsFromProto(ml4, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel3)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel4)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 4, rowCount)
+
+	got, err := ml.GetMarginLevelsByID(ctx, "", "DEADBEEF", entities.Pagination{})
+	assert.NoError(t, err)
+	assert.Equal(t, 2, len(got))
+
+	// We have to truncate the time because Postgres only supports time to microsecond granularity.
+	want1 := *marginLevel3
+	want1.Timestamp = want1.Timestamp.Truncate(time.Microsecond)
+	want1.VegaTime = want1.VegaTime.Truncate(time.Microsecond)
+
+	want2 := *marginLevel4
+	want2.Timestamp = want2.Timestamp.Truncate(time.Microsecond)
+	want2.VegaTime = want2.VegaTime.Truncate(time.Microsecond)
+
+	want := []entities.MarginLevels{want1, want2}
+
+	assert.ElementsMatch(t, want, got)
+}
+
+func testGetMarginLevelsByID(t *testing.T) {
+	testTimeout := time.Second * 10
+	ctx, cancel := context.WithTimeout(context.Background(), testTimeout)
+	defer cancel()
+
+	bs, ml, conn := setupMarginLevelTests(t, ctx)
+
+	var rowCount int
+	err := conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 0, rowCount)
+
+	ml1 := getMarginLevelProto()
+	ml2 := getMarginLevelProto()
+	ml3 := getMarginLevelProto()
+	ml4 := getMarginLevelProto()
+
+	ml2.PartyId = "DEADBAAD"
+
+	ml3.Timestamp = ml2.Timestamp + 1000000000
+	ml3.MaintenanceMargin = "2000"
+	ml3.SearchLevel = "2000"
+
+	ml4.Timestamp = ml2.Timestamp + 1000000000
+	ml4.MaintenanceMargin = "2000"
+	ml4.SearchLevel = "2000"
+	ml4.PartyId = "DEADBAAD"
+
+	block := addTestBlock(t, bs)
+	marginLevel1, err := entities.MarginLevelsFromProto(ml1, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel2, err := entities.MarginLevelsFromProto(ml2, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel1)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel2)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 2, rowCount)
+
+	time.Sleep(time.Second)
+
+	block = addTestBlock(t, bs)
+	marginLevel3, err := entities.MarginLevelsFromProto(ml3, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	marginLevel4, err := entities.MarginLevelsFromProto(ml4, block.VegaTime)
+	require.NoError(t, err, "Converting margin levels proto to database entity")
+
+	err = ml.Upsert(marginLevel3)
+	require.NoError(t, err)
+	err = ml.Upsert(marginLevel4)
+	require.NoError(t, err)
+
+	err = conn.QueryRow(ctx, `select count(*) from margin_levels`).Scan(&rowCount)
+	assert.NoError(t, err)
+	assert.Equal(t, 4, rowCount)
+
+	got, err := ml.GetMarginLevelsByID(ctx, "DEADBEEF", "DEADBEEF", entities.Pagination{})
+	assert.NoError(t, err)
+	assert.Equal(t, 1, len(got))
+
+	// We have to truncate the time because Postgres only supports time to microsecond granularity.
+	want1 := *marginLevel3
+	want1.Timestamp = want1.Timestamp.Truncate(time.Microsecond)
+	want1.VegaTime = want1.VegaTime.Truncate(time.Microsecond)
+
+	want := []entities.MarginLevels{want1}
+
+	assert.ElementsMatch(t, want, got)
+}